+++ CHANGELOG.md
@@ -27,6 +27,7 @@ Compatibility:
 
 * Implemented `-p` CLI option.
 * Implemented "instance variable not initialized" warning.
+* Implemented `-p` and `-a` CLI options.
 
 Performance: